--- backend/.babelrc
@@ -1,21 +1,0 @@
-{
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "useBuiltIns": "usage",
-          "corejs": 3,
-          "targets": {
-            "node": "current"
-          }
-        }
-      ]
-    ],
-    "plugins": [
-      "@babel/plugin-proposal-class-properties",
-      "@babel/plugin-transform-runtime",
-      "rewire"
-    ],
-    "sourceMaps": "inline",
-    "retainLines": true
-}
