+++ backend/package.json
@@ -13,13 +13,9 @@
     "servecmd": "nodemon ./src/server.js",
     "start": "node --max-old-space-size=150 ./src/server.js",
     "lint": "node_modules/.bin/eslint . --no-fix --ignore-pattern 'node_modules'",
-    "lint:fix": "node_modules/.bin/eslint . --fix --ignore-pattern 'node_modules'",
-    "test": "node_modules/.bin/jest --env node",
-    "test-linux": "export NODE_ENV=test && node_modules/.bin/jest",
-    "test:local": "npm run setlocal && npm run test",
-    "test:unit": "npm run settest && npm run test",
+    "test": "NODE_ENV=test vitest run",
+    "test:watch": "vitest",
     "clean": "rm -rf coverage dist",
-    "pretest": "npm run lint",
     "purge": "rm -rf node_modules",
     "reinstall": "npm run purge && npm install",
     "test:unit-linux": "export NODE_ENV=test && npm run test"
@@ -64,56 +60,19 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.5.0",
-    "axios-mock-adapter": "^1.19.0",
-    "babel-jest": "^29.7.0",
-    "babel-plugin-rewire": "^1.2.0",
-    "babel-preset-airbnb": "^5.0.0",
-    "core-js": "^3.12.1",
-    "eslint": "^9.4.0",
-    "jest": "^29.7.0",
-    "jest-sonar-reporter": "^2.0.0",
-    "nodemon": "^3.1.3",
+    "@eslint/js": "^9.9.0",
+    "axios-mock-adapter": "^2.0.0",
+    "core-js": "^3.38.0",
+    "eslint": "^9.9.0",
+    "nodemon": "^3.1.4",
     "redis-mock": "^0.56.3",
-    "supertest": "^7.0.0"
+    "vitest": "^2.0.5"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
     "not ie <= 8"
   ],
-  "jest": {
-    "reporters": [
-      "default",
-      "jest-junit"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json"
-    ],
-    "moduleNameMapper": {
-      "^@/(.*)$": "<rootDir>/src/$1"
-    },
-    "testMatch": [
-      "**/tests/**/*.spec.(js|jsx|ts|tsx)|**/__tests__/*.(js|jsx|ts|tsx)"
-    ],
-    "testURL": "http://localhost",
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "src/**/*.js"
-    ],
-    "testResultsProcessor": "jest-sonar-reporter"
-  },
-  "jest-junit": {
-    "outputDirectory": "./tests/unit/reports/",
-    "outputName": "./junit.xml",
-    "usePathForSuitName": true
-  },
-  "jestSonar": {
-    "reportPath": "coverage",
-    "reportFile": "test-report.xml",
-    "indent": 2
-  },
   "license": "Apache-2.0",
   "repository": {
     "type": "git",
