+++ backend/src/routes/penRequest.js
@@ -13,7 +13,7 @@ import {
   deleteDocument,
   downloadFile,
   uploadFile
-} from '../components/request.js';
+} from '../components/requestHandler.js';
 import { forwardGetReq } from '../components/utils.js';
 import config from '../config/index.js';
 import {
