+++ backend/src/routes/user.js
@@ -1,7 +1,7 @@
 import passport from 'passport';
 import express from 'express';
 
-import { getUserInfo } from '../components/request.js';
+import { getUserInfo } from '../components/requestHandler.js';
 import { isValidBackendToken } from '../components/auth.js';
 
 const router = express.Router();
