--- backend/tests/.eslintrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  env: {
-    jest: true
-  }
-};
